refactor(ui): refactor some component structure
refactor some component structure
