+++ packages/ui/src/containers/TermsOfUsePromiseModal/index.tsx
@@ -1,15 +1,16 @@
 import React, { useContext } from 'react';
 import { create, InstanceProps } from 'react-modal-promise';
 
-import TermsOfUseModal from '@/components/TermsOfUseModal';
 import { PageContext } from '@/hooks/use-page-context';
+
+import TermsOfUseConfirmModal from './TermsOfUseConfirmModal';
 
 const TermsOfUsePromiseModal = ({ isOpen, onResolve, onReject }: InstanceProps<boolean>) => {
   const { setTermsAgreement, experienceSettings } = useContext(PageContext);
   const { termsOfUse } = experienceSettings ?? {};
 
   return (
-    <TermsOfUseModal
+    <TermsOfUseConfirmModal
       isOpen={isOpen}
       termsUrl={termsOfUse?.contentUrl ?? ''}
       onConfirm={() => {
